Split Measurements into two classes
Saver class contains all the saving methods, while the new Measurement class contains all the methods used to run, load instruments, etc.

Should functionally do nothing except organize the code. Need to extensively test before creating PR to master.
