+++ Measurements/heightsweep.py
@@ -3,7 +3,7 @@ import matplotlib.pyplot as plt
 from ..Instruments import piezos, nidaq, montana
 import time, os
 from datetime import datetime
-from ..Utilities.save import Measurement
+from .measurement import Measurement
 from ..Utilities import conversions
 from ..Utilities.utilities import AttrDict
