+++ Measurements/scanspectra.py
@@ -9,7 +9,7 @@ from IPython import display
 from numpy import ma
 from ..Utilities.plotting import plot_mpl
 from ..Instruments import piezos, montana, squidarray
-from ..Utilities.save import Measurement
+from .measurement import Measurement
 from ..Utilities import conversions
 from ..Measurements.daqspectrum import SQUIDSpectrum
 from ..Utilities.utilities import AttrDict
